Got LabVIEW code up and running.
Lesson learned: don't change Reserve's default settings or things will break on you. We are now using the setSetPoint state to set the setpoint to a global variable, where maintainIntakePosition will keep the intake at the proper position with the potentiometer and the elevator motors. We have tested this but the PID values seem drastically off, so we have to work on that as well.  More code still needs to be added - for example in TeleOp and autonomous, but it shouldn't take too long now that the framework itself isn't dying on me.