Added documentation, changed default Setpoint, removed Idle
Added subdiagram labels to each of the states in IntakeController, I removed the idle state bceause Reserve acted as the idle state, and I changed the default setpoint to be 0.5 because now the PID only takes in 0-1. I need to figure those PID inputs out as well as rename the global variable, etc.