DriveForward (drive by time) works!
We tested the DriveForward command as well as the DriveForDistance command. We found that DriveForward works when we used the front panel for defining the number of seconds that the robot should go. For DriveForDistance, we found that the gyro is being weird and seems to drift even when the robot is not turning.